3.4-django-testing/django_testing/requirements.txt
django
djangorestframework
psycopg2-binary
pytest
pytest-django
django-filter